requirements.txt
yt-dlp>=2023.12.30
faster-whisper>=0.10.0
pysrt>=1.1.2
aiohttp>=3.9.0
pyyaml>=6.0.1
requests>=2.31.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
anthropic>=0.8.0
openai>=1.6.0
tqdm>=4.66.0
